Moved all svg icons to images folder
